Moved scTxsCommitmentBuilder to separate file
